requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
buildozer==1.5.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
Cython==3.1.3
distlib==0.4.0
docutils==0.22
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
fonttools==4.59.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.8.0
numpy==1.24.3
packaging==25.0
pandas==2.0.3
passlib==1.7.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plyer==2.1.0
ptyprocess==0.7.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
requests==2.32.5
rsa==4.9.1
sh==2.2.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
